# Conflicts:
#	recipes/certify/all/conanfile.py
#	recipes/continuable/all/conanfile.py
#	recipes/continuable/all/test_package/test_package.cpp
#	recipes/libinterpolate/all/conanfile.py
#	recipes/libpng/all/conanfile.py
#	recipes/r8brain-free-src/all/conanfile.py
#	recipes/rapidcsv/all/conanfile.py
#	recipes/sbp/all/conanfile.py
#	recipes/sbp/all/test_package/CMakeLists.txt
#	recipes/sfml/all/conanfile.py
#	recipes/simde/all/conanfile.py

+++ recipes/sfml/all/conanfile.py
@@ -2,8 +2,9 @@ from conan import ConanFile
 from conan.errors import ConanInvalidConfiguration
 from conan.tools.apple import is_apple_os
 from conan.tools.cmake import CMake, CMakeDeps, CMakeToolchain, cmake_layout
-from conan.tools.files import apply_conandata_patches, export_conandata_patches, get, rmdir, save
+from conan.tools.files import apply_conandata_patches, export_conandata_patches, get, rmdir, save, copy
 from conan.tools.microsoft import is_msvc, is_msvc_static_runtime
+from conan.tools.scm import Version
 import os
 import textwrap
 
@@ -53,6 +54,7 @@ class SfmlConan(ConanFile):
 
     def requirements(self):
         if self.options.window:
+            # FIXME: use cci's glad
             if self.settings.os in ["Windows", "Linux", "FreeBSD", "Macos"]:
                 self.requires("opengl/system")
             if self.settings.os == "Linux":
@@ -62,6 +64,7 @@ class SfmlConan(ConanFile):
             self.requires("freetype/2.13.0")
             self.requires("stb/cci.20220909")
         if self.options.audio:
+            # FIXME: use cci's minimp3
             self.requires("flac/1.4.2")
             self.requires("openal-soft/1.22.2")
             self.requires("vorbis/1.3.7")
@@ -74,14 +77,14 @@ class SfmlConan(ConanFile):
 
     def source(self):
         get(self, **self.conan_data["sources"][self.version], strip_root=True)
-        rmdir(self, os.path.join(self.source_folder, "extlibs"))
+        # sfml/2.6.0 uses minimp3 and glad in extlibs
+        if Version(self.version) < "2.6.0":
+            rmdir(self, os.path.join(self.source_folder, "extlibs"))
 
     def generate(self):
         tc = CMakeToolchain(self)
         tc.cache_variables["SFML_DEPENDENCIES_INSTALL_PREFIX"] = self.package_folder.replace("\\", "/")
-        tc.cache_variables["SFML_MISC_INSTALL_PREFIX"] = os.path.join(
-            self.package_folder, "licenses"
-        ).replace("\\", "/")
+        tc.cache_variables["SFML_MISC_INSTALL_PREFIX"] = os.path.join(self.package_folder, "licenses").replace("\\", "/")
         tc.variables["SFML_BUILD_WINDOW"] = self.options.window
         tc.variables["SFML_BUILD_GRAPHICS"] = self.options.graphics
         tc.variables["SFML_BUILD_NETWORK"] = self.options.network
@@ -89,6 +92,9 @@ class SfmlConan(ConanFile):
         tc.variables["SFML_INSTALL_PKGCONFIG_FILES"] = False
         tc.variables["SFML_GENERATE_PDB"] = False
         tc.variables["SFML_USE_SYSTEM_DEPS"] = True
+        tc.variables["WARNINGS_AS_ERRORS"] = False
+        if Version(self.version) >= "2.6.0":
+            tc.variables["CMAKE_CXX_STANDARD"] = 11
         if is_msvc(self):
             tc.variables["SFML_USE_STATIC_STD_LIBS"] = is_msvc_static_runtime(self)
         tc.generate()
@@ -102,9 +108,12 @@ class SfmlConan(ConanFile):
         cmake.build()
 
     def package(self):
+        if Version(self.version) >= "2.6.0":
+            copy(self, pattern="license.md", dst=os.path.join(self.package_folder, "licenses"), src=self.source_folder)
         cmake = CMake(self)
         cmake.install()
         rmdir(self, os.path.join(self.package_folder, "lib", "cmake"))
+        rmdir(self, os.path.join(self.package_folder, "share"))
 
         # TODO: to remove in conan v2 once cmake_find_package* generators removed
         self._create_cmake_module_alias_targets(
@@ -171,6 +180,9 @@ class SfmlConan(ConanFile):
         def iokit():
             return ["IOKit"] if self.settings.os == "Macos" else []
 
+        def coreservices():
+            return ["CoreServices"] if self.settings.os == "Macos" else []
+
         def coregraphics():
             return ["CoreGraphics"] if self.settings.os == "iOS" else []
 
@@ -200,70 +212,58 @@ class SfmlConan(ConanFile):
                 "target": "sfml-system",
                 "libs": [f"sfml-system{suffix}"],
                 "system_libs": winmm() + pthread() + rt() + android() + log(),
-            }
+            },
         }
         if self.settings.os in ["Windows", "Android", "iOS"]:
             sfml_main_suffix = "-d" if self.settings.build_type == "Debug" else ""
             sfmlmain_libs = [f"sfml-main{sfml_main_suffix}"]
             if self.settings.os == "Android":
                 sfmlmain_libs.append(f"sfml-activity{suffix}")
-            sfml_components.update(
-                {"main": {"target": "sfml-main", "libs": sfmlmain_libs, "system_libs": android() + log()}}
-            )
+            sfml_components.update({
+                "main": {
+                    "target": "sfml-main",
+                    "libs": sfmlmain_libs,
+                    "system_libs": android() + log(),
+                },
+            })
         if self.options.window:
-            sfml_components.update(
-                {
-                    "window": {
-                        "target": "sfml-window",
-                        "libs": [f"sfml-window{suffix}"],
-                        "requires": ["system"] + opengl() + xorg() + libudev(),
-                        "system_libs": gdi32() + winmm() + usbhid() + android() + opengles_android(),
-                        "frameworks": (
-                            foundation()
-                            + appkit()
-                            + iokit()
-                            + carbon()
-                            + uikit()
-                            + coregraphics()
-                            + quartzcore()
-                            + coremotion()
-                            + opengles_ios()
-                        ),
-                    }
-                }
-            )
+            sfml_components.update({
+                "window": {
+                    "target": "sfml-window",
+                    "libs": [f"sfml-window{suffix}"],
+                    "requires": ["system"] + opengl() + xorg() + libudev(),
+                    "system_libs": gdi32() + winmm() + usbhid() + android() + opengles_android(),
+                    "frameworks": foundation() + appkit() + iokit() + carbon() +
+                                  uikit() + coregraphics() + quartzcore() +
+                                  coreservices() + coremotion() + opengles_ios(),
+                },
+            })
         if self.options.graphics:
-            sfml_components.update(
-                {
-                    "graphics": {
-                        "target": "sfml-graphics",
-                        "libs": [f"sfml-graphics{suffix}"],
-                        "requires": ["window", "freetype::freetype", "stb::stb"],
-                    }
-                }
-            )
+            sfml_components.update({
+                "graphics": {
+                    "target": "sfml-graphics",
+                    "libs": [f"sfml-graphics{suffix}"],
+                    "requires": ["window", "freetype::freetype", "stb::stb"],
+                },
+            })
         if self.options.network:
-            sfml_components.update(
-                {
-                    "network": {
-                        "target": "sfml-network",
-                        "libs": [f"sfml-network{suffix}"],
-                        "requires": ["system"],
-                        "system_libs": ws2_32(),
-                    }
-                }
-            )
+            sfml_components.update({
+                "network": {
+                    "target": "sfml-network",
+                    "libs": [f"sfml-network{suffix}"],
+                    "requires": ["system"],
+                    "system_libs": ws2_32(),
+                },
+            })
         if self.options.audio:
-            sfml_components.update(
-                {
-                    "audio": {
-                        "target": "sfml-audio",
-                        "libs": [f"sfml-audio{suffix}"],
-                        "requires": ["system", "flac::flac", "openal-soft::openal-soft", "vorbis::vorbis"],
-                        "system_libs": android(),
-                    }
-                }
-            )
+            sfml_components.update({
+                "audio": {
+                    "target": "sfml-audio",
+                    "libs": [f"sfml-audio{suffix}"],
+                    "requires": ["system", "flac::flac", "openal-soft::openal-soft", "vorbis::vorbis"],
+                    "system_libs": android(),
+                },
+            })
 
         return sfml_components
 
@@ -294,12 +294,8 @@ class SfmlConan(ConanFile):
                 self.cpp_info.components[component].frameworks = frameworks
 
                 # TODO: to remove in conan v2 once cmake_find_package* generators removed
-                self.cpp_info.components[component].build_modules["cmake_find_package"] = [
-                    self._module_file_rel_path
-                ]
-                self.cpp_info.components[component].build_modules["cmake_find_package_multi"] = [
-                    self._module_file_rel_path
-                ]
+                self.cpp_info.components[component].build_modules["cmake_find_package"] = [self._module_file_rel_path]
+                self.cpp_info.components[component].build_modules["cmake_find_package_multi"] = [self._module_file_rel_path]
 
         _register_components(self._sfml_components)
